Removed conversion to INT from Exercise67

+++ Learning Python/Chapter 06/Exercise67.py
@@ -2,6 +2,6 @@ def is_power(a,b):
     if a==b:
         return True
     if a%b == 0:
-        return is_power(int(a/b), b)
+        return is_power(a/b, b)
     else:
         return False
